requirements.txt
awscrt==0.16.10
awsiot==0.1.3
awsiotsdk==1.12.5
Printrun==2.0.0rc7
python-decouple==3.8
simplejson==3.18.3
